+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.6 - 2021-09-24
 - Exclude `resourcepaths` table from db backup
 - Ignore mysql import errors 🤞🏻
+- Better error output
 
 ## 1.0.5 - 2021-05-12
 - Update craft-auto-migrate to use `project-config/apply`